gfx/tests/mochitest/test_bug513439.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=513439
-->
<head>
  <title>Test for Bug 513439</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=513439">Mozilla Bug 513439</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 513439 **/

var prefService = SpecialPowers.Cc["@mozilla.org/preferences-service;1"]
                               .getService(SpecialPowers.Ci.nsIPrefService);
var layoutCSSBranch = prefService.getBranch("layout.css.");
var oldVal = layoutCSSBranch.getCharPref("devPixelsPerPx");

try {
  var domWindowUtils = window.QueryInterface(SpecialPowers.Ci.nsIInterfaceRequestor)
                             .getInterface(SpecialPowers.Ci.nsIDOMWindowUtils);
  var devPxPerCSSPx = domWindowUtils.screenPixelsPerCSSPixel;

  layoutCSSBranch.setCharPref("devPixelsPerPx", "2");
  is(domWindowUtils.screenPixelsPerCSSPixel, 2, "devPixelsPerPx wasn't set correctly");

  layoutCSSBranch.setCharPref("devPixelsPerPx", "1.5");
  is(domWindowUtils.screenPixelsPerCSSPixel, 1.5, "devPixelsPerPx wasn't set correctly");

} finally {
  layoutCSSBranch.setCharPref("devPixelsPerPx", oldVal);
}

</script>
</pre>
</body>
</html>
